+++ src/lin_alloc.c
@@ -3,7 +3,7 @@
 #include <stdint.h>
 #include "ualloc.h"
 
-int8_t lin_alloc_init(size_t size, linear_allocator * allocator){
+int lin_alloc_init(size_t size, linear_allocator * allocator){
     allocator->buffer = malloc(size);
     if (!allocator->buffer){
         *allocator = (linear_allocator) {NULL, NULL, 0};
@@ -14,7 +14,7 @@ int8_t lin_alloc_init(size_t size, linear_allocator * allocator){
     return 0;
 }
 
-int8_t lin_alloc(size_t size, linear_allocator * allocator, void ** buffer){
+int lin_alloc(size_t size, linear_allocator * allocator, void ** buffer){
     uint8_t * pos = (uint8_t *)allocator->position;
     uint8_t * buff = (uint8_t *)allocator->buffer;
     if (pos + size > (buff+allocator->size)){
